requirements.txt
Django==5.0.1
djangorestframework==3.14.0
drf-yasg==1.21.7
pandas==2.1.4
openpyxl==3.1.2
simplekml==1.3.6
requests==2.31.0
polyline==2.0.2
python-decouple==3.8
gunicorn==21.2.0
psycopg2-binary==2.9.9
